ProductDetails open on item click of WishList & Cart items

+++ app/src/main/java/com/bs/ecommerce/product/model/data/WishListItem.kt
@@ -21,7 +21,7 @@ data class WishListItem(
     @SerializedName("Picture")
     val picture: PictureModel?,
     @SerializedName("ProductId")
-    val productId: Int?,
+    val productId: Long?,
     @SerializedName("ProductName")
     val productName: String?,
     @SerializedName("ProductSeName")
